Comenzamos con la clase Pasajero

+++ bdViajes.sql
@@ -1,27 +1,27 @@
 CREATE DATABASE bdviajes; 
 
 CREATE TABLE empresa(
-    idempresa bigint AUTO_INCREMENT,
+    idempresa INT AUTO_INCREMENT,
     enombre varchar(150),
     edireccion varchar(150),
     PRIMARY KEY (idempresa)
     ) ENGINE=InnoDB DEFAULT CHARSET=utf8 AUTO_INCREMENT=1;
 
 CREATE TABLE responsable (
-    rnumeroempleado bigint AUTO_INCREMENT,
-    rnumerolicencia bigint,
+    rnumeroempleado INT AUTO_INCREMENT,
+    rnumerolicencia INT,
 	rnombre varchar(150), 
     rapellido  varchar(150), 
     PRIMARY KEY (rnumeroempleado)
     )ENGINE=InnoDB DEFAULT CHARSET=utf8 AUTO_INCREMENT=1;;
 	
 CREATE TABLE viaje (
-    idviaje bigint AUTO_INCREMENT,
+    idviaje INT AUTO_INCREMENT,
 	vdestino varchar(150),
     vcantmaxpasajeros int,
     rdocumento varchar(15),
-	idempresa bigint,
-    rnumeroempleado bigint,
+	idempresa INT,
+    rnumeroempleado INT,
     vimporte float,
     tipoAsiento varchar(150), /*primera clase o no, semicama o cama*/
     idayvuelta varchar(150), /*si no*/
@@ -37,7 +37,7 @@ CREATE TABLE pasajero (
     pnombre varchar(150), 
     papellido varchar(150), 
 	ptelefono int, 
-	idviaje bigint,
+	idviaje INT,
     PRIMARY KEY (rdocumento),
 	FOREIGN KEY (idviaje) REFERENCES viaje (idviaje)	
     )ENGINE=InnoDB DEFAULT CHARSET=utf8;
